amaya_env/requirements.txt
streamlit==1.4.0
openai==0.28.0
sqlalchemy
psycopg2-binary  # For PostgreSQL
requests
instaloader
toml
sentence_transformers
pinecone
pytesseract   
Pillow
pinecone-client